+++ src/contracts/Token.sol
@@ -6,39 +6,39 @@ pragma solidity >=0.4.22 <0.9.0;
 contract Token {
 	string public name = 'EMG';
 	string public symbol = 'EMG';
-	uint256 public decimals = 18;
-	uint256 public totalSupply;
-	mapping(address => uint256) public balanceOf;
+	uint public decimals = 18;
+	uint public totalSupply;
+	mapping(address => uint) public balanceOf;
 	// address maps to exchange which maps to address and amount
-	mapping(address => mapping(address => uint256)) public allowance;
+	mapping(address => mapping(address => uint)) public allowance;
 
-	event Transfer(address indexed from, address indexed to, uint256 value);
-	event Approval(address indexed owner, address indexed spender, uint256 value);
+	event Transfer(address indexed from, address indexed to, uint value);
+	event Approval(address indexed owner, address indexed spender, uint value);
 
 	constructor() public {
 		totalSupply = 100 * (10**decimals);
 		balanceOf[msg.sender] = totalSupply;
 	}
 	
-	function _transferWithEmit(address _from, address _to, uint256 _value) internal {
+	function _transferWithEmit(address _from, address _to, uint _value) internal {
 		balanceOf[_from] = balanceOf[_from] - _value;
 		balanceOf[_to] = balanceOf[_to] + _value;
 		emit Transfer(_from, _to, _value);
 	}
 
-	function transfer(address _to, uint256 _value) public returns (bool success) {
+	function transfer(address _to, uint _value) public returns (bool success) {
 		require(balanceOf[msg.sender] >= _value);
 		_transferWithEmit(msg.sender, _to, _value);
 		return true;
 	}
 	
-	function approve(address _spender, uint256 _value) public returns (bool success) {
+	function approve(address _spender, uint _value) public returns (bool success) {
 		allowance[msg.sender][_spender] = _value;
 		emit Approval(msg.sender, _spender, _value);
 		return true;
 	}
 	
-	function transferFrom(address _from, address _to, uint256 _value) public returns (bool success) {	
+	function transferFrom(address _from, address _to, uint _value) public returns (bool success) {	
 		require(_value <= balanceOf[_from]);	
 		require(_value <= allowance[_from][msg.sender]);	
 		_transferWithEmit(_from, _to, _value);
